Added fetch for getting all financial accounts, and updated the component to show
